refactor(web): extract the clinical bolus table so the report page keeps legal exports
